guestbook_project/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
djangorestframework==3.12.4